requirements.txt
psycopg2-binary==2.9.9
python-telegram-bot==21.6
python-dotenv==1.0.1
sqlalchemy==2.0.30
requests==2.32.3